refactor(logic): move file system package
