+++ x/logic/fs/object.go
@@ -13,6 +13,9 @@ type Object struct {
 	modTime time.Time
 }
 
+var _ fs.File = (*Object)(nil)
+var _ fs.FileInfo = (*Object)(nil)
+
 func NewObject(src []byte, uri *url.URL, modTime time.Time) Object {
 	return Object{
 		reader:  bytes.NewReader(src),
