+++ x/logic/fs/wasm.go
