+++ x/logic/fs/wasm_test.go
